# backend/order-service/doc/swagger.yaml
openapi: 3.0.0
info:
  title: Order Service API
  version: v1
  description: API for managing orders.
paths:
...
  /api/order/create/:
    post:
      summary: Create a new order
      description: Creates a new order with the provided details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request (e.g., invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the order.
        ownerId:
          type: integer
          description: ID of the order owner.
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
      required:
        - id
        - ownerId
        - order_items
    OrderItem:
      type: object
      properties:
        id_food_item:
          type: integer
          description: ID of the food item.
        quantity:
          type: integer
          description: Quantity of the food item.
        name:
          type: string
          description: Name of the food item.
        img:
          type: string
          description: URL of the food item image.
      required:
        - id_food_item
        - quantity
    CreateOrderRequest:
      type: object
      properties:
        ownerId:
          type: integer
          description: ID of the order owner.
        foodItems:
          type: array
          items:
            $ref: '#/components/schemas/FoodItem'
      required:
        - ownerId
        - foodItems
    FoodItem:
      type: object
      properties:
        id:
          type: integer
          description: ID of the food item.
        quantity:
          type: integer
          description: Quantity of the food item.
      required:
        - id
        - quantity
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message.